add mainboard.py
